fix: separating views from serializers
The notifications app had views and serializers mixed in a single
serializers.py file, with only a comment indicating where a new "backend/notifications/views.py" file was supposed  have been created.
This also broke imports which referred to a file which was never created.

This appears to be an artifact of AI code
